+++ src/test/resources/org/pgcodekeeper/core/it/diff/ms/compare_ms_statistics_original.sql
@@ -9,3 +9,6 @@ CREATE STATISTICS stat1 ON [dbo].[t6] (col1, col2);
 
 CREATE STATISTICS stat2 ON [dbo].[t6] (col1, col2, col3)
 WITH FULLSCAN;
+
+CREATE STATISTICS stat3 ON [dbo].[t6] (col1, col2, col3)
+WITH FULLSCAN;