+++ src/test/resources/org/pgcodekeeper/core/it/diff/ms/modify_ms_table_with_partition_new.sql
@@ -6,3 +6,12 @@ CREATE TABLE [dbo].[table_partition_01](
     [id] [int] NULL
 ) ON [ps_range_right]([id])
 GO
+
+CREATE TABLE [dbo].[table_partition_02](
+    [id] [int] NULL
+) ON [ps_range_left]([id])
+GO
+
+ALTER TABLE [dbo].[table_partition_02]
+	ADD CONSTRAINT [part_name] PRIMARY KEY CLUSTERED ([id]) ON [ps_range_left]([id])
+GO