+++ src/test/resources/org/pgcodekeeper/core/it/jdbc/ms/ms_sys_ver_table.sql
@@ -7,8 +7,8 @@ CREATE TABLE [dbo].[hist_t1](
     [DepartmentName] [varchar] (50) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [ManagerID] [int] NULL,
     [ParentDepartmentNumber] [char] COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
-    [ValidFrom] [datetime2] NOT NULL,
-    [ValidTo] [datetime2] NOT NULL,
+    [ValidFrom] [datetime2] (7) NOT NULL,
+    [ValidTo] [datetime2] (7) NOT NULL,
     [col1] [int] NULL
 ) ON [PRIMARY]
 WITH (DATA_COMPRESSION = PAGE)
@@ -28,8 +28,8 @@ CREATE TABLE [dbo].[t1](
     [DepartmentName] [varchar] (50) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [ManagerID] [int] NULL,
     [ParentDepartmentNumber] [char] COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
-    [ValidFrom] [datetime2] GENERATED ALWAYS AS ROW START HIDDEN NOT NULL,
-    [ValidTo] [datetime2] GENERATED ALWAYS AS ROW END HIDDEN NOT NULL,
+    [ValidFrom] [datetime2] (7) GENERATED ALWAYS AS ROW START HIDDEN NOT NULL,
+    [ValidTo] [datetime2] (7) GENERATED ALWAYS AS ROW END HIDDEN NOT NULL,
     [col1] [int] NULL,
     PERIOD FOR SYSTEM_TIME ([ValidFrom], [ValidTo])
 ) ON [PRIMARY]